Upgraded Git package manager scripts

+++ getdir.php
@@ -5,27 +5,30 @@ $pkg = (isset($_REQUEST['pkg'])) ? $_REQUEST['pkg'] : 'from';
 $repo = (isset($_REQUEST['repo'])) ? $_REQUEST['repo'] : 'system';
 $branch = (isset($_REQUEST['branch'])) ? $_REQUEST['branch'] : '';
 $user = (isset($_REQUEST['user'])) ? $_REQUEST['user'] : 'infofintech';
-if ($action == 'i' || $action == 's' || $action == 'o') {
-    if ($pkg == "from" && $repo != "" && $user != "") {
-        $hostArr = explode('://', $host);
-        $fsockopen = fsockopen($hostArr[1], 80, $errno, $errstr, 10);
-        if ($fsockopen != false) {
-            if (file_exists($repo)) {
-                exec('chmod -R 777 .'); exec('rm -rf '.$repo);
-            } if ($branch != '') {
-                exec('git clone -b '.$branch.' '.$host.'/'.$user.'/'.$repo);
-            } else {
-                exec('git clone '.$host.'/'.$user.'/'.$repo);
-            } chmod($repo, 0777); exec('chmod -R 777 .');
+if (!file_exists('get.lock')) {
+    if ($action == 'i' || $action == 's' || $action == 'o') {
+        if ($pkg == "from" && $repo != "" && $user != "") {
+            $hostArr = explode('://', $host);
+            file_put_contents('get.lock', ''); chmod('get.lock', 0777);
+            $fsockopen = fsockopen($hostArr[1], 80, $errno, $errstr, 10);
+            if ($fsockopen != false) {
+                if (file_exists($repo)) {
+                    exec('chmod -R 777 .'); exec('rm -rf '.$repo);
+                } if ($branch != '') {
+                    exec('git clone -b '.$branch.' '.$host.'/'.$user.'/'.$repo);
+                } else {
+                    exec('git clone '.$host.'/'.$user.'/'.$repo);
+                } chmod($repo, 0777); exec('chmod -R 777 .');
+            } chmod('get.lock', 0777); unlink('get.lock');
         }
-    }
-} elseif ($action == 'r' || $action == 'p' || $action == 'm') {
-    if ($pkg != "" && $repo != "" && $user != "") {
-        if (file_exists($pkg)) { exec('chmod -R 777 .'); exec('rm -rf '.$pkg); }
-        header('Location: getdir.php?key=i&host='.$host.'&pkg=from&repo='.$repo.'&branch='.$branch.'&user='.$user);
-    }
-} elseif ($action == 'd' || $action == 'u' || $action == 'x') {
-    if ($pkg != "" && $repo == 'from' && $user == 'here') {
-        if (file_exists($pkg)) { exec('chmod -R 777 .'); exec('rm -rf '.$pkg); }
+    } elseif ($action == 'r' || $action == 'p' || $action == 'm') {
+        if ($pkg != "" && $repo != "" && $user != "") {
+            if (file_exists($pkg)) { exec('chmod -R 777 .'); exec('rm -rf '.$pkg); }
+            header('Location: getdir.php?key=i&host='.$host.'&pkg=from&repo='.$repo.'&branch='.$branch.'&user='.$user);
+        }
+    } elseif ($action == 'd' || $action == 'u' || $action == 'x') {
+        if ($pkg != "" && $repo == 'from' && $user == 'here') {
+            if (file_exists($pkg)) { exec('chmod -R 777 .'); exec('rm -rf '.$pkg); }
+        }
     }
 }
